requirements.txt
fastapi
uvicorn
pillow
tensorflow